Use event stream instead of callback for repr
