+++ pyproject.toml
@@ -1,2 +1,3 @@
 [tool.black]
 line-length = 100
+extend-exclude = "test_notebook.ipynb"
